requirements.txt
fastapi
dependency_injector
sqlalchemy
bcrypt
pydantic
python-jose
pandas
numpy
yfinance
plotly
